Graceful thread shutdown on unexpected errors
In the event of an unexpected error in a thread, this change ensures that all other
threads that are still running will get a graceful shutdown period of 30
seconds (like monit). After that, the remaining threads are killed and the
threaded worker process will fail with an exception.
This ensures that monit can restart the failed process.

DelayedJobs already handles all possible errors during job execution,
so the graceful shutdown is only applied to errors that occur
after a job has finished.

Co-authored-by: Johannes Haass <johannes.haass@sap.com>